Moved all medium specific files from compute/properties/impl to medium/<dim>/<med>/<prop>
