+++ src/compute/compute_properties.cpp
@@ -17,19 +17,19 @@ specfem::compute::properties::properties(
       nspec, mapping, tags, h_medium_tags, h_property_tags, n_elastic_isotropic,
       n_elastic_anisotropic, n_acoustic);
 
-  acoustic_isotropic = specfem::compute::impl::properties::material_properties<
+  acoustic_isotropic = specfem::medium::material_properties<
       specfem::element::medium_tag::acoustic,
       specfem::element::property_tag::isotropic>(
       nspec, n_acoustic, ngllz, ngllx, mapping, tags, materials,
       h_property_index_mapping);
 
-  elastic_isotropic = specfem::compute::impl::properties::material_properties<
+  elastic_isotropic = specfem::medium::material_properties<
       specfem::element::medium_tag::elastic,
       specfem::element::property_tag::isotropic>(
       nspec, n_elastic_isotropic, ngllz, ngllx, mapping, tags, materials,
       h_property_index_mapping);
 
-  elastic_anisotropic = specfem::compute::impl::properties::material_properties<
+  elastic_anisotropic = specfem::medium::material_properties<
       specfem::element::medium_tag::elastic,
       specfem::element::property_tag::anisotropic>(
       nspec, n_elastic_anisotropic, ngllz, ngllx, mapping, tags, materials,
