+++ src/compute/impl/value_containers.cpp
@@ -61,4 +61,4 @@ void specfem::compute::impl::compute_number_of_elements_per_medium(
 //     specfem::compute::impl::kernels::material_kernels>;
 
 // template class specfem::compute::impl::value_containers<
-//     specfem::compute::impl::properties::material_properties>;
+//     specfem::medium::material_properties>;
