--- src/domain/impl/elements/elastic/elastic2d.cpp
@@ -1,4 +1,0 @@
-#include "domain/impl/elements/elastic/elastic2d.hpp"
-#include "domain/impl/elements/elastic/elastic2d.tpp"
-
-// Explicit function template instantiation
